+++ CHANGELOG.md
@@ -13,11 +13,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added game environment switcher
 - Added game edition switcher
 - Added changelog to updated components toast
+- Added wine tools to settings
+
+### Fixed
+
+- Added whitespaces removing from environment values
 
 ### Changed
 
 - Improved game repairing feature
 - Replaced `curl` dependency by native code
+- Replaced static image by spinner in wine / dxvk version selection
+- Made wine / dxvk versions always visible if they're downloaded
 
 ## [3.3.0] - 24.03.2023